autogen/requirements.txt
autogen-agentchat>=0.4.5
autogen-ext[openai]
autogenstudio
openai-agents
fastmcp
